fix: User-controlled data in numeric cast [#1987] (#1996)
* fix: User-controlled data in numeric cast [#1987]

+++ hugegraph-api/src/main/java/com/baidu/hugegraph/api/traversers/AdamicAdarAPI.java
@@ -67,7 +67,7 @@ public class AdamicAdarAPI extends API {
                       @QueryParam("max_degree")
                       @DefaultValue(DEFAULT_MAX_DEGREE) long maxDegree,
                       @QueryParam("limit")
-                      @DefaultValue(DEFAULT_ELEMENTS_LIMIT) long limit) {
+                      @DefaultValue(DEFAULT_ELEMENTS_LIMIT) int limit) {
         LOG.debug("Graph [{}] get adamic adar between '{}' and '{}' with " +
                   "direction {}, edge label {}, max degree '{}' and limit '{}'",
                   graph, current, other, direction, edgeLabel, maxDegree,
